# Backend/movie-app-go/docs/swagger.yaml
basePath: /api/v1
definitions:
  models.AdminReviewRequest:
    properties:
      admin_review:
        type: string
    type: object
  models.Genre:
    properties:
      genre_id:
        type: string
      genre_name:
        maxLength: 100
        minLength: 2
        type: string
    required:
    - genre_id
    - genre_name
    type: object
...
  models.Movie:
    properties:
      admin_review:
        type: string
      description:
        maxLength: 5000
        minLength: 10
        type: string
      genres:
        items:
          $ref: '#/definitions/models.Genre'
        type: array
      id:
        type: string
      imdb_id:
        type: string
      poster_url:
        type: string
      ranking:
        $ref: '#/definitions/models.Ranking'
      release_year:
        maximum: 2100
        minimum: 1888
        type: integer
      title:
        maxLength: 200
        minLength: 2
        type: string
      youtube_id:
        type: string
    required:
    - description
    - genres
    - imdb_id
    - poster_url
    - ranking
    - release_year
    - title
    - youtube_id
    type: object
  models.Ranking:
    properties:
      ranking_name:
        maxLength: 100
        minLength: 2
        type: string
      ranking_value:
        type: integer
    required:
    - ranking_name
    - ranking_value
    type: object
...
host: localhost:5000
info:
  contact: {}
  description: API for the movie app.
  title: Movie API
  version: "1.0"
paths:
  /addmovie:
    post:
      consumes:
      - application/json
      parameters:
      - description: Movie
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.Movie'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Add a movie
      tags:
      - movies
...
  /genres:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get all genres
      tags:
      - movies
  /movie/{imdbId}:
    get:
      parameters:
      - description: IMDb ID
        in: path
        name: imdbId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get movie by IMDb id
      tags:
      - movies
  /movie/review/{imdbId}:
    patch:
      consumes:
      - application/json
      parameters:
      - description: IMDb ID
        in: path
        name: imdbId
        required: true
        type: string
      - description: Admin review
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.AdminReviewRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Update admin review for a movie
      tags:
      - movies
  /movies:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: List movies
      tags:
      - movies
  /recommendatedmovies:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get recommended movies
      tags:
      - movies
  /recommendations-ai:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get AI recommended movies
      tags:
      - movies
  /searchmovies:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Search movies
      tags:
      - movies
schemes:
- http
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"